requirements.txt
argcomplete==2.0.0
boto3==1.26.27
botocore==1.29.27
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==2.1.1
click==8.1.3
durationpy==0.5
et-xmlfile==1.1.0
Flask==2.2.2
hjson==3.1.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
numpy==1.23.5
openai==0.25.0
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
zappa==0.56.1
zipp==3.11.0

python-dotenv~=0.19.2